// lib/anty/animation/types.ts
export type EmotionType =
  | 'happy'
  | 'sad'
  | 'angry'
  | 'shocked'
  | 'excited'    // Level 4 positive - jump + spin (no confetti)
  | 'celebrate'  // Level 5 positive - EPIC (confetti!)
  | 'pleased'    // Level 2 positive - bounce + happy eyes
  | 'spin'
  | 'jump'
  | 'idea'
  | 'back-forth'
  | 'look-around'
  | 'wink'
  | 'nod'
  | 'headshake'
  | 'look-left'
  | 'look-right'
  | 'super'
  | 'smize';